requirements.txt
flask
oci
numpy
scikit-learn
joblib
python-dotenv